Fix/#57 - 게시판 수정 (#58)
* #23 - 회원가입 에러 메시지 수정

* #23 - 회원가입 표시 수정

* #23 - 회원가입 표시 수정2

* #23 - 휴대폰 번호 글자수 제한

* #26 - 게시글 작성 시 지역명이 나오게 수정

* #26 - delete_yn 을 enum 으로 수정

* #23 - 휴대폰 번호를 숫자만 입력 가능하게 만듦

* #57 - 댓글 수정 기능 구현

* #57 - 질문 등록 기능 수정

* #57 - 댓글 기능 비로그인시 작성 불가, 댓글 기능 권한 체크, 사용자id 체크

+++ src/main/java/com/example/hellpyending/article/controller/ArticleCommentController.java
@@ -5,7 +5,11 @@ import com.example.hellpyending.article.domain.ArticleComment;
 import com.example.hellpyending.article.form.ArticleCommentForm;
 import com.example.hellpyending.article.service.ArticleCommentService;
 import com.example.hellpyending.article.service.ArticleService;
+import com.example.hellpyending.user.UserService;
+import com.example.hellpyending.user.entity.Users;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -13,8 +17,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
+import java.security.Principal;
 
 @RequestMapping("/articleComment")
 @Controller
@@ -23,8 +29,11 @@ public class ArticleCommentController {
     private final ArticleService articleService;
     private final ArticleCommentService articleCommentService;
 
+    private final UserService userService;
+
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create/{id}")
-    public String detail(Model model, @PathVariable long id, @Valid ArticleCommentForm articleCommentForm, BindingResult bindingResult) {
+    public String detail(Principal principal, Model model, @PathVariable long id, @Valid ArticleCommentForm articleCommentForm, BindingResult bindingResult) {
         Article article = this.articleService.getArticle(id);
 
         if ( bindingResult.hasErrors() ) {
@@ -32,15 +41,54 @@ public class ArticleCommentController {
             return "article_detail";
         }
 
-        articleCommentService.create(article, articleCommentForm.getContent());
+        Users users = userService.getUser(principal.getName());
+
+        articleCommentService.create(article, articleCommentForm.getContent(),users);
 
         return "redirect:/article/detail/%d".formatted(id);
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String articleCommentDelete(@PathVariable("id") long id) {
+    public String articleCommentDelete(Principal principal, @PathVariable("id") long id) {
         ArticleComment articleComment = this.articleCommentService.getArticleComment(id);
+
+        if (!articleComment.getUsers().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
+
         this.articleCommentService.delete(articleComment);
         return String.format("redirect:/article/detail/%s", articleComment.getArticle().getId());
+    }
+
+    @PreAuthorize("isAuthenticated()")
+    @GetMapping("/modify/{id}")
+    public String answerModify(ArticleCommentForm articleCommentForm, @PathVariable("id") Long id, Principal principal) {
+        ArticleComment articleComment = articleCommentService.getArticleComment(id);
+
+        if (!articleComment.getUsers().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
+
+        articleCommentForm.setContent(articleComment.getComment());
+
+        return "articleComment_form";
+    }
+    @PreAuthorize("isAuthenticated()")
+    @PostMapping("/modify/{id}")
+    public String articleCommentModify(@Valid ArticleCommentForm articleCommentForm, BindingResult bindingResult, @PathVariable("id") Long id, Principal principal) {
+        if (bindingResult.hasErrors()) {
+            return "article_detail";
+        }
+
+        ArticleComment articleComment = articleCommentService.getArticleComment(id);
+
+        if (!articleComment.getUsers().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
+
+        articleCommentService.modify(articleComment, articleCommentForm.getContent());
+
+        return "redirect:/article/detail/%d".formatted(articleComment.getArticle().getId(), articleComment.getId());
     }
 }
