+++ src/main/java/com/example/hellpyending/article/controller/ArticleController.java
@@ -1,6 +1,7 @@
 package com.example.hellpyending.article.controller;
 
 import com.example.hellpyending.article.domain.Article;
+import com.example.hellpyending.article.exception.DataNotFoundException;
 import com.example.hellpyending.article.form.ArticleCommentForm;
 import com.example.hellpyending.article.form.ArticleForm;
 import com.example.hellpyending.article.service.ArticleService;
@@ -9,11 +10,13 @@ import com.example.hellpyending.user.UserService;
 import com.example.hellpyending.user.entity.Users;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.security.Principal;
@@ -69,32 +72,49 @@ public class ArticleController {
         if (bindingResult.hasErrors()) {
             return "article_form";
         }
+
+
         Users users = this.userService.getUser(principal.getName());
-        articleService.create(articleForm.getTitle(), articleForm.getContent(), users.getAddress_1st());
+        articleService.create(articleForm.getTitle(), articleForm.getContent(), users.getAddress_1st(), users);
         return "redirect:/article/list"; // 질문 저장 후 질문 목록으로 이동
     }
-
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/modify/{id}")
-    public String articleModify(ArticleForm articleForm, @PathVariable(name = "id") Long id) {
+    public String articleModify(ArticleForm articleForm, @PathVariable(name = "id") Long id, Principal principal) {
 
         Article article = this.articleService.getArticle(id);
 
-        articleForm.setId(article.getId());
+        if (article == null) {
+            throw new DataNotFoundException("%d번 질문은 존재하지 않습니다.");
+        }
+
+        if (!article.getUsers().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
+
         articleForm.setTitle(article.getTitle());
         articleForm.setContent(article.getContent());
         articleForm.setUpdate(article.getUpdate());
         articleForm.setAreaName(article.getAreaName());
 
-        return "articleUpdate_form";
+        return "article_form";
     }
 
-    @PostMapping("/modify")
-    public String articleModify(Model model, @Valid ArticleForm articleForm, BindingResult bindingResult) {
+    @PostMapping("/modify/{id}")
+    public String articleModify(Model model, @Valid ArticleForm articleForm, BindingResult bindingResult, Principal principal, @PathVariable("id") Integer id) {
         if (bindingResult.hasErrors()) {
-            return "question_form";
+            return "article_form";
         }
 
-        System.out.println("articleForm.getId()  " + articleForm.getId());
+        Article article = this.articleService.getArticle(id);
+
+        if (article == null) {
+            throw new DataNotFoundException("%d번 질문은 존재하지 않습니다.");
+        }
+
+        if (!article.getUsers().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
 
         boolean result = articleService.modify(articleForm.getId(), articleForm.getTitle(), articleForm.getContent(), articleForm.getAreaName());
 
@@ -108,8 +128,14 @@ public class ArticleController {
 
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String articleDelete(@PathVariable(name = "id") Long id) {
+    public String articleDelete(Principal principal, @PathVariable(name = "id") Long id) {
+        Article article = articleService.getArticle(id);
+
+        if (!article.getUsers().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
         boolean result = articleService.delete(id);
         if (result){
             return "redirect:/article/list";
